+++ crates/runner/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [0.186.7](https://github.com/vm0-ai/vm0/compare/runner-rs-v0.186.6...runner-rs-v0.186.7) (2026-09-06)
+
 ## [0.186.6](https://github.com/vm0-ai/vm0/compare/runner-rs-v0.186.5...runner-rs-v0.186.6) (2026-09-06)
 
 ## [0.186.5](https://github.com/vm0-ai/vm0/compare/runner-rs-v0.186.4...runner-rs-v0.186.5) (2026-09-05)